Issue #SB-13986 fix: fetching frameworkdata
