+++ src/app/client/src/app/modules/program/interfaces/index.ts
@@ -1,1 +1,2 @@
+export * from './program';
 export * from './programHeader';
